+++ CHANGES.rst
@@ -53,6 +53,10 @@ Compute
   (GITHUB-1481)
   [Eis D. Zaster - @Eis-D-Z]
 
+- [OpenStack] Enable to get Quota Set detail. 
+  (GITHUB-1495)
+  [Miguel Caballer - @micafer]
+
 Changes in Apache Libcloud 3.2.0
 --------------------------------